refactor: ChatInput Refactor ''
