+++ client/src/Component/Molecules/index.tsx
@@ -3,7 +3,7 @@ export * from "./MainPage/MainHeaderCenter";
 export * from "./MainPage/MainBodyLeftButtonContainer";
 export * from "./MainPage/MainBodyRight.ButtonContainer";
 
-export * from "./Chat/ChatInput";
+export * from "./Chat/ChatInput/ChatInput";
 export * from "./Chat/ChatProfileInfoContainer";
 export * from "./Chat/ChatRoomBasic";
 export * from "./Chat/ChatRoomFooter";
