+++ client/src/Component/Page/Page/UserPage.tsx
@@ -1,5 +1,5 @@
 import { Routes, Route, Navigate } from "react-router-dom";
-import { CowDogPage, TeamSettingPage, MyPage, ChatListPage, RequestPage } from "..";
+import { CowDogPage, TeamSettingPage, MyPage, ChatListPage, RequestPage, ChatRoom } from "..";
 
 export const UserPage = () => {
   return (
@@ -9,7 +9,7 @@ export const UserPage = () => {
       <Route path="/mypage/*" element={<MyPage />} />
       <Route path="/chatList" element={<ChatListPage />} />
       <Route path="/Request" element={<RequestPage />} />
-      {/* <Route path="/ChatRoom" element={<ChatRoom />} /> */}
+      <Route path="/ChatRoom" element={<ChatRoom />} />
       <Route path="*" element={<Navigate to="/" replace />} />
     </Routes>
   );
