+++ client/src/Component/Template/ChatRoomLeft/ChatRoomLeft.tsx
@@ -1,4 +1,4 @@
-import { ChatInput } from "@Molecules/Chat/ChatInput";
+import { ChatInput } from "@Molecules/Chat/ChatInput/ChatInput";
 import { ChatDetail } from "@Organism/Chat/ChatDetail/ChatDetail";
 
 export const ChatRoomLeft = () => {
